Migrate clients-kt/main module to kafka 3.6.0

+++ clients-kt/src/main/kotlin/org/apache/kafka/common/record/DefaultRecordsSend.kt
@@ -28,9 +28,9 @@ class DefaultRecordsSend<T : TransferableRecords>(
     @Throws(IOException::class)
     override fun writeTo(
         channel: TransferableChannel,
-        previouslyWritten: Long,
+        previouslyWritten: Int,
         remaining: Int
-    ): Long = records().writeTo(
+    ): Int = records().writeTo(
         channel = channel,
         position = previouslyWritten,
         length = remaining
